Application now uses webview based repo for generating media samples
Removed unused youtube code and dependencies
