+++ app/build.gradle
@@ -50,20 +50,11 @@ dependencies {
 
     implementation 'com.google.android.exoplayer:exoplayer:2.10.6'
     implementation 'com.google.android.exoplayer:extension-mediasession:2.10.6'
-    implementation 'com.github.HaarigerHarald:android-youtubeExtractor:1.9.0'
 
     implementation 'com.github.bumptech.glide:glide:4.10.0'
 
     implementation 'org.koin:koin-android:2.0.1'
     implementation 'org.koin:koin-android-scope:2.0.1'
-
-    implementation 'com.squareup.okhttp3:okhttp:4.2.2'
-
-    implementation 'com.squareup.retrofit2:retrofit:2.6.1'
-
-    implementation 'com.squareup.moshi:moshi:1.9.1'
-    implementation 'com.squareup.moshi:moshi-kotlin:1.9.1'
-    kapt 'com.squareup.moshi:moshi-kotlin-codegen:1.9.1'
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.2"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.2"
