--- app/src/main/java/com/jzallas/backdrop/extensions/moshi/Moshi.kt
@@ -1,7 +1,0 @@
-package com.jzallas.backdrop.extensions.moshi
-
-import com.squareup.moshi.Moshi
-
-inline fun <reified T> Moshi.deserialize(json: String): T? {
-  return adapter(T::class.java).fromJson(json)
-}
