--- app/src/main/java/com/jzallas/backdrop/youtube/repository/YtFileRepository.kt
@@ -1,21 +1,0 @@
-package com.jzallas.backdrop.youtube.repository
-
-import android.util.SparseArray
-import at.huber.youtubeExtractor.YtFile
-import com.jzallas.backdrop.youtube.extractor.YouTubeExtractor
-import com.jzallas.backdrop.youtube.api.YouTubeUrlParser
-import kotlinx.coroutines.runBlocking
-
-class YtFileRepository(
-  private val extractor: YouTubeExtractor,
-  private val urlParser: YouTubeUrlParser
-) {
-
-  fun getYtFiles(url: String): SparseArray<YtFile> = runBlocking {
-    extractor.extract(
-      url = urlParser.parse(url),
-      parseDashManifest = true,
-      includeWebM = true
-    )
-  } ?: throw IllegalStateException("Unable to fetch ytFiles.")
-}
